refactor!: restructure directories
BREAKING CHANGE: import paths must be updated
