+++ ablm_eval/__init__.py
@@ -1,4 +1,3 @@
-from .configs import *
 from .tasks import *
 from .utils import *
 from .run import *
