--- ablm_eval/configs/__init__.py
@@ -1,6 +1,0 @@
-from .classification import *
-from .inference import *
-from .moe_routing import *
-from .mutation_analysis import *
-from .naturalness import *
-from .per_position_inference import *
