--- ablm_eval/plots/__init__.py
@@ -1,5 +1,0 @@
-from .mutation_analysis import *
-from .naturalness import *
-from .per_position import *
-from .routing import *
-from .tables import *
