+++ ablm_eval/tasks/__init__.py
@@ -1,6 +1,6 @@
 from .classification import *
 from .inference import *
 from .moe_routing import *
-from .mutation_analysis import *
-from .naturalness import *
+from .mutation_prediction import *
+from .naturalness_prediction import *
 from .per_position_inference import *
