+++ ablm_eval/utils/__init__.py
@@ -1,6 +1,6 @@
+from ..tasks.compare_registry import *
 from .data import *
-from .compare import *
+from .directories import *
 from .load_model import *
 from .metrics import *
-from .output import *
-from .process_datasets import *
+from .tables import *
