+++ test_script.py
@@ -1,5 +1,14 @@
-from ablm_eval.configs import *
-from ablm_eval import evaluate_ablms, compare_results, compare_task
+from ablm_eval import (
+    InferenceConfig,
+    PerPositionConfig,
+    MutationPredConfig,
+    ClassificationConfig,
+    RoutingConfig,
+    NaturalnessConfig,
+    evaluate_ablms,
+    compare_results,
+    compare_task
+)
 
 
 def main():
